Updated Interface
The interface for interacting with the shapes has been overhaled to centralise the access and improve reliability. The lines now show the exponents on them if the exponent is greater than 1. Just need to add options for factors not being added multiple times and a few other ease of use things.