Prime Check Field
The primes can now be checked using an input field. Also it now works on the laptop. The scope of the integers has been upgraded and rationalised to the type: superint which is just an unsigned int. The project needs to be adapted so that it can use static Qt installs.